pyproject.toml
[project]
name = "rag-prac1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.6",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.26",
    "langchain-groq>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
    "unstructured>=0.22.16",
]
